feat: expose `parse_dimension()`
