+++ src/index.ts
@@ -6,6 +6,7 @@ export { parse_selector } from './parse-selector'
 export { parse_atrule_prelude } from './parse-atrule-prelude'
 export { parse_declaration } from './parse-declaration'
 export { parse_value } from './parse-value'
+export { parse_dimension } from './parse-dimension'
 export { tokenize } from './tokenize'
 export { walk, traverse, SKIP, BREAK } from './walk'
 export { is_custom, is_vendor_prefixed, str_equals, str_starts_with, str_index_of } from './string-utils'
